separate client and server code
