+++ lib/grpc_web/message_framing.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'grpc_web/grpc_web_response'
-require 'grpc_web/grpc_web_request'
 require 'grpc_web/message_frame'
 
 module GRPCWeb
@@ -10,20 +8,6 @@ module GRPCWeb
   # byte length of the frame body.
   module MessageFraming
     class << self
-      def unframe_request(request)
-        frames = unframe_content(request.body)
-        ::GRPCWeb::GRPCWebRequest.new(
-          request.service, request.service_method, request.content_type, request.accept, frames,
-        )
-      end
-
-      def frame_response(response)
-        framed = response.body.map do |frame|
-          frame_content(frame.body, frame.frame_type)
-        end.join
-        ::GRPCWeb::GRPCWebResponse.new(response.content_type, framed)
-      end
-
       def frame_content(content, frame_type = ::GRPCWeb::MessageFrame::PAYLOAD_FRAME_TYPE)
         length_bytes = [content.bytesize].pack('N')
         "#{frame_type.chr}#{length_bytes}#{content}"
