+++ lib/grpc_web/server/error_callback.rb
