+++ lib/grpc_web/server/service_class_validator.rb
