+++ lib/grpc_web/server/text_coder.rb
