+++ spec/integration/envoy_server_ruby_client_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-require 'grpc_web/client'
+require 'grpc_web/client/client'
 require 'hello_services_pb'
 require 'test_grpc_server'
 require 'test_hello_service'
